Implement signature verification in service provider
If the "xmlsec" feature is activated, ensure that the signatures of the
IdP's SAML response are valid. While doing so, remove any elements that
are neither signed nor have any children that are signed, protecting
against the potential exploit of an unsigned assertion placed among
signed assertions being interpreted as valid.