Make status in responses optional
We are already only checking the StatusCode if it has a value. So the
following was already accepted:
  <samlp:Status><samlp:StatusCode/></samlp:Status>

This commit also accepts the entire Status node not existing in the
response.
This occurs if only the Assertion in a SAMLResponse is signed and the
reduce_to_signed_xml function removes the unsigned status element.
If this occurs the reduced response will not pass the serialization
roundtrip due to the now missing Status node being required.