Add required NameID field to LogoutRequest
According to the specification the `<LogoutRequest>` element is required
to contain `<saml:BaseID>` or `<saml:NameID>` or `<saml:EncryptedID>`.
This commit adds support for `<saml:NameID>` being parsed optionally to
still parse messages containing one of the other two successfully.

Support for the other two wasn't added since they are also not supported
for the `<Subject>` element which would allow them as well.

This could have been combined with the `SubjectNameID` struct but their
serialization implemented later will be different so this was added as a
separate struct.