detect static or dynamic API at build time
The libxmlsec1 library is available with two similar but incompatible
APIs.  On Debian and Ubuntu systems, the library is built with the
--disable-crypto-dl and --disable-apps-crypto-dl configure options which
results in OpenSSL-specific symbols like xmlSecOpenSSLAppKeyLoad().  On
every other platform I could find, these options are not used, and the
library can dynamically select a backend using symbols like
xmlSecCryptoAppKeyLoad().

The only way this can be detected is examining the CFLAGS we get from
xmlsec1-config.  The XMLSEC_CRYPTO_DYNAMIC_LOADING macro does not appear
to be included in any shipped header file.  The build script now checks
for this definition and uses conditional compilation to substitute
either the OpenSSL-specific static symbols, or the dynamic symbols.