handle --enable-size-t in XmlSecKey::from_memory
xmlsec enabled --enable-size-t by default in 1.3.0, which changes the
xmlSecSize typedef to size_t. This changes the `dataSize` parameter in
`xmlSecOpenSSLAppKeyLoadMemory` either `usize` or `u32`, depending on
the version of xmlsec and the compiler options used by the distribution
we happen to build this crate on.

Instead of casting to u32, we can ask the compiler to choose the right
conversion for us with `TryInto`, and panic if the conversion would
truncate.