Do not send NameIDPolicy with a blank Format
Previously, if the `name_id_format` function returned None, this would
cause `NameIdPolicy` to have None as a format, leading to the following
tag in an AuthnRequest:

```xml
<saml2p:NameIDPolicy AllowCreate="true"/>
```

Skip creating a `Some(NameIDPolicy)` if `name_id_format` returns `None`.

Additionally, do not override if `authn_name_id_format` is set to
`Unspecified`: previously this would cause `name_id_format` to return
`None` where a caller explicitly set it to `Unspecified`.